fix(frontend): extract pipeline progress utils for react-refresh lint
Move resolveStepStatuses and countCompletedSteps to pipelineProgress.utils.ts so PipelineProgress.tsx only exports components (eslint react-refresh/only-export-components).

Made-with: Cursor
