+++ manu-gen/frontend/src/features/jobs/components/JobDetailPage.tsx
@@ -8,7 +8,8 @@ import { usePipeline } from "../../pipelines/hooks/usePipeline";
 import { JobJourneyChart } from "./JobJourneyChart";
 import { JobHistory } from "./JobHistory";
 import { JobDetailKpis } from "./JobDetailKpis";
-import { PipelineProgress, countCompletedSteps } from "./PipelineProgress";
+import { PipelineProgress } from "./PipelineProgress";
+import { countCompletedSteps } from "./pipelineProgress.utils";
 import { QrPreview } from "./QrPreview";
 import { JobStatusBadge } from "./JobStatusBadge";
 import { useJobHistory } from "../hooks/useJobHistory";
